ctlmgr: Allow explicitly specifying host filter
This is useful for hosts with multiple IP addresses, where all
of them are on the master's subnet as well. In a typical such
setup, it used to be up to chance (or rather the order of
OS-level routes) which of them would be used for the local
socket endpoint, and consequently the host filter.

Another fix would be to change this to filter against all local
IP addresses instead of just one.